+++ lib/setalpha.dart
@@ -1,11 +1,7 @@
 // Copyright (c) 2017, Richard Ambler. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library setalpha;
 
-export 'src/setalpha_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+part 'src/color.dart';
+part 'src/setalpha.dart';