+++ lib/src/setalpha_base.dart
@@ -1,7 +1,0 @@
-// Copyright (c) 2017, Richard Ambler. All rights reserved. Use of this source code
-// is governed by a BSD-style license that can be found in the LICENSE file.
-
-library setalpha;
-
-part 'color.dart';
-part 'setalpha.dart';