+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ## 0.5.0
 
 - Set the alpha level of CSS colors.
-- Supports color name, hex values, rgb (and rgba) expressions as input.
+- Supports color name, hex values, rgb (and rgba) expressions as input to the function `setAlpha`.
 
 ## 0.6.0
 
-- Also supports hsl expressions as input.
+Now also supports hsl expressions as input to function `setAlpha`.
+
+## 0.7.0
+
+Added the function `Color.nearest`, which returns the css color nearest in rgb space to input rgb values.